requirements.txt
# requirements.txt
fastapi
uvicorn
supabase
openai
python-dotenv
httpx
soundfile
numpy
TTS
sounddevice
